Added user-specific events and protected routes

+++ server/routes/eventRoutes.js
@@ -1,11 +1,15 @@
 const express = require("express");
 const router = express.Router();
 const Event = require("../models/Event");
+const authMiddleware = require("../middleware/authMiddleware");
 
 // CREATE EVENT
-router.post("/", async (req, res) => {
+router.post("/", authMiddleware, async (req, res) => {
   try {
-    const event = new Event(req.body);
+    const event = new Event({
+  ...req.body,
+  user: req.user.id,
+});
 
     await event.save();
 
@@ -19,23 +23,41 @@ router.post("/", async (req, res) => {
 });
 
 // GET ALL EVENTS
-router.get("/", async (req, res) => {
+router.get("/", authMiddleware, async (req, res) => {
   try {
 
-    const events = await Event.find();
+    const events = await Event.find({
+      user: req.user.id,
+    });
 
     res.json(events);
 
   } catch (error) {
+
     res.status(500).json({
       error: error.message,
     });
+
   }
 });
 
 // DELETE EVENT
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", authMiddleware, async (req, res) => {
   try {
+
+    const event = await Event.findById(req.params.id);
+
+    if (!event) {
+      return res.status(404).json({
+        message: "Event not found",
+      });
+    }
+
+    if (event.user.toString() !== req.user.id) {
+      return res.status(403).json({
+        message: "Unauthorized",
+      });
+    }
 
     await Event.findByIdAndDelete(req.params.id);
 
@@ -44,15 +66,31 @@ router.delete("/:id", async (req, res) => {
     });
 
   } catch (error) {
+
     res.status(500).json({
       error: error.message,
     });
+
   }
 });
 
 // UPDATE EVENT
-router.put("/:id", async (req, res) => {
+router.put("/:id", authMiddleware, async (req, res) => {
   try {
+
+    const event = await Event.findById(req.params.id);
+
+    if (!event) {
+      return res.status(404).json({
+        message: "Event not found",
+      });
+    }
+
+    if (event.user.toString() !== req.user.id) {
+      return res.status(403).json({
+        message: "Unauthorized",
+      });
+    }
 
     const updatedEvent = await Event.findByIdAndUpdate(
       req.params.id,
@@ -63,9 +101,11 @@ router.put("/:id", async (req, res) => {
     res.json(updatedEvent);
 
   } catch (error) {
+
     res.status(500).json({
       error: error.message,
     });
+
   }
 });
